\name{factorize}
\alias{factorize}
\title{Complete Prime Factorization}
\description{Given an integer m (or an object that can be coerced to an integer), which is not prime and is two or greater, returns a matrix with two columns. The first column gives the factors and the second column gives the multiplicities.}
\usage{
factorize (m)
}
\arguments{
\item{m}{An integer as described above.}
}
\examples{
#completely factorize the number 229075
factorize (229075)
}
