\name{pnf}
\alias{pnf}
\title{Integer Factorization}
\description{Given a positive integer m (or an object that can be coerced to an integer), returns NA if m is one or a prime number, and returns the smallest integer factor (excluding one) otherwise.}
\usage{
pnf (m)
}
\arguments{
\item{m}{A positive integer.}
}
\examples{
#find the smallest factor of the number 187
pnf (187)
}
